refactor(tests): test/bugs move to other categories
