--- src/test/bugs/bugs.tact
@@ -1,7 +1,0 @@
-import "./issue42.tact";
-import "./issue43.tact";
-import "./issue53.tact";
-import "./issue74.tact";
-import "./issue117.tact";
-import "./issue211.tact";
-import "./large-contract.tact";
