--- src/test/bugs/issue53.tact
@@ -1,6 +1,0 @@
-contract Issue53 {
-    init() {}
-    receive() {
-        require(1 == 2, "condition can`t be...");
-    }
-}
