+++ src/test/codegen/all-contracts.tact
@@ -1,0 +1,3 @@
+import "./empty-message.tact";
+import "./large-contract.tact";
+import "./struct-field-storage-annotation.tact";
