+++ src/test/codegen/empty-message.tact
