+++ src/types/test/scope-loops.tact
